Ensure version is stamped during publishing
When publishing, the ldflag we were passing to stamp in a version was
pointing at James' fork, leading to these warning when trying to use
the plugin:

```
warning: resource plugin vsphere is expected to have version
>=0.16.1, but has ; the wrong version may be on your path, or this may
be a bug in the plugin
```

In an ideal world, we would not need to hard code the import path of
the package, and that's something we can probably solve by using `go
list ./pkg/version` to solve across all repositories.

I opened an overall tracking issue to address the above.

Fixes #13